More work towards stopping tests from depending on extension code
