+++ src/test/flutter_web_create_tests/flutter_web_create.test.ts
@@ -2,10 +2,10 @@ import * as assert from "assert";
 import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
-import { InternalExtensionApi } from "../../extension/extension";
 import { internalApiSymbol } from "../../extension/symbols";
 import { fsPath } from "../../extension/utils";
 import { dartCodeExtensionIdentifier } from "../../shared/constants";
+import { InternalExtensionApi } from "../../shared/interfaces";
 import { sb, waitForResult } from "../helpers";
 import sinon = require("sinon");
